==> 0. NQ loader.ipynb <==
This is the code with which the "NQ dev ds with answers" has been generated.

resetDb()

# Wikipedia Parser
Parse and structures the NQ-dev dataset Wikipedia pages to create the graph that will be used as the KB of the experiment.

from utils.wikipediaParser import *
import io
from tqdm import tqdm
import pandas as pd
import config.EnvLoader as el

resetCsv()

parsed_urls = []
parsed_titles = []
no_answer = []


with open("v1.0-simplified_nq-dev-all.jsonl") as file:
    for line in tqdm(file):
        line = io.StringIO(line)
        jsonObj = pd.read_json(path_or_buf=line, lines=True)
        # If the document has already been parsed, skip
        if jsonObj['document_url'][0] in parsed_urls or jsonObj['document_title'][0] in parsed_titles:
            continue
        # If the document does not have an answer, skip
        ann_lst = [ann['long_answer']['start_byte'] for ann in jsonObj['annotations'][0]]
        if len(set(ann_lst)) == 1 and ann_lst[0] == -1:
            no_answer.append(jsonObj["question_text"][0])
            continue
        parse(jsonObj.iloc[0])
        parsed_urls.append(jsonObj["document_url"][0])
        parsed_titles.append(jsonObj["document_title"][0])

len(no_answer)

Move csvs into db's import folder

createConstrains()

query_lst = neo4jExecutor()

linkChunks()

Move to DBpedia DMBS and run

extractRelations(query_lst)
extractLabels(query_lst)

Copy DBpedia csvs and move back to the db

addDBpediaInfo()

# Extract answers
URI = "neo4j://localhost"
AUTH = ("neo4j", el.NEO4J_PWD)
result = []

with open("v1.0-simplified_nq-dev-all.jsonl") as file:
    for line in tqdm(file):
        line = io.StringIO(line)
        jsonObj = pd.read_json(path_or_buf=line, lines=True)
        result.append(extractAnswers(jsonObj, URI, AUTH))

df_result = pd.DataFrame(result, columns = ['question', 'result']) 
print("Pages not found in Neo4j: " + str(len(df_result[df_result["result"] == "Page not found in Neo4j"])))
print("Questions with no answers: " + str(len(df_result[df_result["result"] == "No answer provided"])))
print("Non-found chunks: " + str(len(df_result[df_result["result"] == "not found"])))
print("Found answers: " + str(len(df_result[df_result["result"] == "found"])))


# Import answers
writeAnswers(URI, AUTH)

# Remove non-requested DBpedia relaitonships
with GraphDatabase.driver(URI, auth=AUTH) as driver:
    pair_lst, summary, keys = driver.execute_query(
        """CALL apoc.periodic.iterate(
        "MATCH (a:Page)-[r:vector_rel]-(b:Page)
        WHERE NOT (a)-[:cited_in]->(:Chunk)-[:same_page]->(:Chunk)-[:cites]->(b)
        AND NOT (a)-[:has_chunk]->(:Chunk)-[:cites]->(b)
        AND NOT (b)-[:has_chunk]->(:Chunk)-[:cites]->(a)
        AND NOT (a)-[:cited_in]->(:Chunk)-[:cites]->(b)
        AND ID(a) > ID(b)
        RETURN a, r, b",
        "DELETE r",
        {batchSize:1000, parallel:false}
        )
        YIELD batches, total, timeTaken, committedOperations
        RETURN batches, total, timeTaken, committedOperations;""",
    )

pair_lst[0]['committedOperations']
